organizing parser
