+++ src/core/parser/ast/ast.c
@@ -6,11 +6,11 @@
 /*   By: valero <valero@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/25 23:03:17 by valero            #+#    #+#             */
-/*   Updated: 2025/11/27 00:32:47 by valero           ###   ########.fr       */
+/*   Updated: 2025/11/30 21:49:27 by valero           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "parser_internal.h"
+#include "ast_internal.h"
 
 static void	*ft_destroy_ast_node(
 				t_ast_node **self_ref,
@@ -37,10 +37,8 @@ t_ast	*ft_create_ast(t_lexer *lexer)
 	return (ast);
 }
 
-t_ast_node	*ft_create_ast_node(
-				t_token **tokens,
-				t_ast_node_type type,
-				void *exec)
+t_ast_node	*ft_create_ast_node(t_token **tokens,
+				t_ast_node_type type, void *exec)
 {
 	t_ast_node	*node;
 	int			tokens_size;
@@ -71,7 +69,8 @@ static void	*ft_destroy_ast_node(
 	self = *self_ref;
 	if (self->tokens)
 		free(self->tokens);
-	free_exec(self->exec);
+	if (free_exec)
+		free_exec(self->exec);
 	free(self);
 	*self_ref = NULL;
 	return (NULL);
