+++ src/core/parser/ast/ast_internal.h
@@ -1,16 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parser.h                                           :+:      :+:    :+:   */
+/*   ast_internal.h                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: valero <valero@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/11/05 00:14:45 by valero            #+#    #+#             */
-/*   Updated: 2025/11/30 20:40:11 by valero           ###   ########.fr       */
+/*   Created: 2025/11/30 20:30:55 by valero            #+#    #+#             */
+/*   Updated: 2025/11/30 20:33:45 by valero           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef PARSER_H
-# define PARSER_H
+#ifndef AST_INTERNAL_H
+# define AST_INTERNAL_H
+
+# include "ast.h"
+
+void	ft_print_ast(t_ast *self);
 
 #endif
