+++ src/core/parser/ast_build/ast_build.h
@@ -1,16 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parser.h                                           :+:      :+:    :+:   */
+/*   ast_build.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: valero <valero@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/11/05 00:14:45 by valero            #+#    #+#             */
-/*   Updated: 2025/11/30 20:40:11 by valero           ###   ########.fr       */
+/*   Created: 2025/11/30 20:37:28 by valero            #+#    #+#             */
+/*   Updated: 2025/11/30 21:44:14 by valero           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef PARSER_H
-# define PARSER_H
+#ifndef AST_BUILD_H
+# define AST_BUILD_H
+
+# include "../ast/ast.h"
+
+t_ast	*ft_ast_build(t_lexer *lexer, void *exec);
 
 #endif
