+++ src/core/parser/ast_build/ast_build_utils.c
@@ -1,34 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   sintax.c                                           :+:      :+:    :+:   */
+/*   ast_build_utils.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: brunofer <brunofer@student.42.fr>          +#+  +:+       +#+        */
+/*   By: valero <valero@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/11/25 21:51:10 by valero            #+#    #+#             */
-/*   Updated: 2025/11/27 15:14:44 by brunofer         ###   ########.fr       */
+/*   Created: 2025/11/30 20:54:04 by valero            #+#    #+#             */
+/*   Updated: 2025/11/30 21:10:40 by valero           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "parser_internal.h"
+#include "ast_build_internal.h"
 
+static int	ft_get_precedence(t_token *token);
 static void	ft_jump_to_paren_closing(t_token **tokens, int *idx);
-
-int	ft_get_precedence(t_token *token)
-{
-	int	max_precedence;
-
-	max_precedence = 4;
-	if (token->type == TOKEN_PAREN_OPEN)
-		return (max_precedence - 0);
-	if (ft_is_redirect_node(token))
-		return (max_precedence - 1);
-	if (token->type == TOKEN_PIPE)
-		return (max_precedence - 2);
-	if (token->type == TOKEN_AND || token->type == TOKEN_OR)
-		return (max_precedence - 3);
-	return (0);
-}
 
 int	ft_find_lower_precedence(t_token **tokens)
 {
@@ -56,6 +41,48 @@ int	ft_find_lower_precedence(t_token **tokens)
 	return (lower_precedence_idx);
 }
 
+static int	ft_get_precedence(t_token *token)
+{
+	int	max_precedence;
+
+	max_precedence = 4;
+	if (token->type == TOKEN_PAREN_OPEN)
+		return (max_precedence - 0);
+	if (ft_is_redirect_node(token))
+		return (max_precedence - 1);
+	if (token->type == TOKEN_PIPE)
+		return (max_precedence - 2);
+	if (token->type == TOKEN_AND || token->type == TOKEN_OR)
+		return (max_precedence - 3);
+	return (0);
+}
+
+t_binary_tree_node	*ft_get_next_node(t_ast *ast,
+						t_binary_tree_node *tree_node, bool on_left)
+{
+	if (!tree_node)
+		return (ast->tree->root);
+	else
+	{
+		if (on_left)
+			return (tree_node->left);
+		return (tree_node->right);
+	}
+}
+
+t_buid_ast_context	ft_create_buid_ast_context(t_ast *ast,
+						t_binary_tree_node *tree_node, bool on_left,
+						t_lexer_subset *prev_subset)
+{
+	t_buid_ast_context	context;
+
+	context.ast = ast;
+	context.tree_node = tree_node;
+	context.on_left = on_left;
+	context.prev_subset = prev_subset;
+	return (context);
+}
+
 static void	ft_jump_to_paren_closing(t_token **tokens, int *idx)
 {
 	int	opened_parens;
@@ -70,14 +97,4 @@ static void	ft_jump_to_paren_closing(t_token **tokens, int *idx)
 		if (tokens[*idx]->type == TOKEN_PAREN_CLOSE)
 			opened_parens--;
 	}
-}
-
-bool	ft_is_redirect_node(t_token *token)
-{
-	if (token->type == TOKEN_REDIRECT_OUT
-		|| token->type == TOKEN_REDIRECT_IN
-		|| token->type == TOKEN_HERE_DOC_IN
-		|| token->type == TOKEN_APPEND_OUT)
-		return (true);
-	return (false);
 }
